% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_region}
\alias{add_region}
\title{Add Region}
\usage{
add_region(data, lat_col = "latitude", specific_region = NA)
}
\arguments{
\item{data}{dataframe object to modify}

\item{specific_region}{optional argument. If you want to return only those observations in a specific region, type it as a string here.}
}
\value{
a dataframe identical to the first except with an added region column (and filtered if specific_region is specified)
}
\description{
Add Region
}
\examples{
#' temp = array(data = 1:1000000, dim = c(10,10,10000))
lon = seq(from = .5, to = 359.5, by = .5)
lat = seq(from = 89.5, to = -89.5, by = -1)
ok = array_to_df(temp, lons = lon, lats = lat)
add_region(ok, lat_col = 'y', specific_region = 'Temperate')
}
