% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{array_to_df}
\alias{array_to_df}
\title{Array to Dataframe}
\usage{
array_to_df(array, lons, lats, start_date = date("1850-01-01"))
}
\arguments{
\item{array}{input array}

\item{lons}{vector of longitudes that correspond to the slices of the array}

\item{lats}{vector of latitudes that correspond to the slices of the array}
}
\value{
a dataframe with an added column representing the values of the input array, and with the columns x and y representing the mapped location of that value
}
\description{
Array to Dataframe
}
\examples{
temp = array(data = 1:1000000, dim = c(10,10,10000))
lon = seq(from = .5, to = 359.5, by = .5)
lat = seq(from = 89.5, to = -89.5, by = -1)
array_to_df(temp, lons = lon, lats = lat)
}
