% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GetRegions}
\alias{GetRegions}
\title{Get Regions}
\usage{
GetRegions(df, region)
}
\arguments{
\item{df}{data to manipulate}

\item{region}{desired region}
}
\value{
filtered dataframe
}
\description{
Get Regions
}
\examples{
temp = array(data = 1:1000000, dim = c(10,10,10000))
lon = seq(from = .5, to = 359.5, by = .5)
lat = seq(from = 89.5, to = -89.5, by = -1)
ok = array_to_df(temp, lons = lon, lats = lat)
names(ok) = c('temp', 'date', 'longitude', 'latitude')
GetRegions(ok, region = 'Tropics')
}
