% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{slice_data}
\alias{slice_data}
\title{Slice Data}
\usage{
slice_data(my.array, years = c(1900, 2000), origin = 1850)
}
\arguments{
\item{years}{a two element vector specifying the start and end dates for the data that you are interested in}

\item{origin}{the first year of the array, in int form}

\item{array}{the array to slice}
}
\value{
an array containing only the data for the dates specified
}
\description{
Slice Data
}
\examples{
temp = array(data = 1:1000000, dim = c(10,10,10000))
slice_data(temp)
}
