% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{year_to_date}
\alias{year_to_date}
\title{Year to Date}
\usage{
year_to_date(number)
}
\arguments{
\item{number}{the year (as an integer) that you want as a date compatible string}
}
\value{
a string representing the first day of January in the year you specified
}
\description{
Year to Date
}
\examples{
year_to_date(1950)
}
